separate querystring test
